requirements.txt
google-generativeai
python-dotenv
pandas
matplotlib
google-cloud-bigquery